test: split large regression modules by concern
Extract provenance-specific argv and digest tests into their
own module and move SBOM temp-artifact and vulnerability
cleanup checks into a dedicated test file.

Trim the original regression files down to their remaining
concerns so the suite is easier to navigate and future
refactors can land with more targeted test ownership.

Refresh a few stale assertions while preserving coverage of
the current behavior under the split layout.
