+++ tests/test_provenance_subjects.py
@@ -1,0 +1,470 @@
+#!/usr/bin/env python3
+# SPDX-FileCopyrightText: 2026 Technology Innovation Institute (TII)
+#
+# SPDX-License-Identifier: Apache-2.0
+# pylint: disable=missing-function-docstring,protected-access
+
+"""Focused tests for provenance digest and subject handling."""
+
+import json
+import logging
+from types import SimpleNamespace
+
+from provenance import main as provenance_main
+
+
+def test_get_dependencies_supports_nix_2_33_wrapped_json(monkeypatch):
+    drv_path = "/nix/store/0aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-root.drv"
+    dep_basename = "1bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb-dependency.drv"
+    dep_path = f"/nix/store/{dep_basename}"
+    digest = "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"
+    real_exec_cmd = provenance_main.exec_cmd
+
+    def fake_exec_cmd(cmd, **kwargs):
+        if cmd == [
+            "nix-store",
+            "--query",
+            "--references",
+            "--include-outputs",
+            drv_path,
+        ]:
+            return SimpleNamespace(stdout=f"{dep_path}\n")
+        if cmd[:4] == ["nix", "derivation", "show", "-r"]:
+            return SimpleNamespace(
+                stdout=json.dumps(
+                    {
+                        "derivations": {
+                            dep_basename: {
+                                "name": "dependency",
+                                "env": {"version": "1.2.3"},
+                            }
+                        },
+                        "version": 4,
+                    }
+                )
+            )
+        if cmd[:3] == ["nix", "hash", "convert"]:
+            return real_exec_cmd(cmd, **kwargs)
+        raise AssertionError(f"unexpected command: {cmd} kwargs={kwargs}")
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fake_exec_cmd)
+    monkeypatch.setattr(
+        provenance_main,
+        "query_store_hashes",
+        lambda paths: ["sha256:1b8m03r63zqhnjf7l5wnldhh7c134ap5vpj0850ymkq1iyzicy5s"],
+    )
+
+    assert provenance_main.get_dependencies(drv_path) == [
+        {
+            "name": "dependency",
+            "uri": dep_path,
+            "digest": {"sha256": digest},
+            "annotations": {"version": "1.2.3"},
+        }
+    ]
+
+
+def test_normalize_digest_does_not_shell_out(monkeypatch):
+    def fail_exec_cmd(cmd, **kwargs):
+        raise AssertionError(f"unexpected command: {cmd} kwargs={kwargs}")
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fail_exec_cmd)
+
+    assert provenance_main._normalize_digest(
+        "sha256:1b8m03r63zqhnjf7l5wnldhh7c134ap5vpj0850ymkq1iyzicy5s"
+    ) == {"sha256": "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"}
+    assert provenance_main._normalize_digest(
+        "sha256-ungWv48Bz+pBQUDeXa4iI7ADYaOWF3qctBD/YfIAFa0="
+    ) == {"sha256": "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"}
+    assert provenance_main._normalize_digest(
+        "77a94a83ccab42a68278ac5d3e340dcefecd736dd4feff1de71dec137b6b44ce",
+        "r:sha256",
+    ) == {"sha256": "77a94a83ccab42a68278ac5d3e340dcefecd736dd4feff1de71dec137b6b44ce"}
+
+
+def test_normalize_digest_rejects_overflowing_nix32_values(monkeypatch):
+    def fail_exec_cmd(cmd, **kwargs):
+        raise AssertionError(f"unexpected command: {cmd} kwargs={kwargs}")
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fail_exec_cmd)
+
+    assert provenance_main._normalize_digest("sha256:" + ("z" * 52)) is None
+
+
+def test_dependency_package_logs_non_normalized_digest_fallback(caplog):
+    drv_path = "/nix/store/1bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb-dependency.drv"
+
+    with caplog.at_level(logging.WARNING, logger=provenance_main.LOG.name):
+        package = provenance_main._dependency_package(drv_path, "sha999:abc", {}, {})
+
+    assert package == {
+        "name": "dependency",
+        "uri": drv_path,
+        "digest": {"sha999": "abc"},
+    }
+    assert "Falling back to non-normalized digest" in caplog.text
+
+
+def test_get_dependencies_prefers_fixed_output_digest_for_output_paths(monkeypatch):
+    drv_path = "/nix/store/0aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-root.drv"
+    dep_drv_basename = "1bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb-source.drv"
+    dep_out_basename = "2ccccccccccccccccccccccccccccccc-source"
+    dep_out_path = f"/nix/store/{dep_out_basename}"
+    metadata_digest = "77a94a83ccab42a68278ac5d3e340dcefecd736dd4feff1de71dec137b6b44ce"
+    real_exec_cmd = provenance_main.exec_cmd
+
+    def fake_exec_cmd(cmd, **kwargs):
+        if cmd == [
+            "nix-store",
+            "--query",
+            "--references",
+            "--include-outputs",
+            drv_path,
+        ]:
+            return SimpleNamespace(stdout=f"{dep_out_path}\n")
+        if cmd[:4] == ["nix", "derivation", "show", "-r"]:
+            return SimpleNamespace(
+                stdout=json.dumps(
+                    {
+                        "derivations": {
+                            dep_drv_basename: {
+                                "name": "source",
+                                "outputs": {
+                                    "out": {
+                                        "path": dep_out_basename,
+                                        "hash": metadata_digest,
+                                        "hashAlgo": "r:sha256",
+                                    }
+                                },
+                                "env": {"version": "1.2.3"},
+                            }
+                        },
+                        "version": 4,
+                    }
+                )
+            )
+        if cmd[:3] == ["nix", "hash", "convert"]:
+            return real_exec_cmd(cmd, **kwargs)
+        raise AssertionError(f"unexpected command: {cmd} kwargs={kwargs}")
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fake_exec_cmd)
+    monkeypatch.setattr(
+        provenance_main,
+        "query_store_hashes",
+        lambda paths: ["sha256:09i0w2qz3i5yp7m3yziq4z2n2r2v9s6d3n8j4x1q8k0m5a6b7c8d"],
+    )
+
+    assert provenance_main.get_dependencies(drv_path) == [
+        {
+            "name": "source",
+            "uri": dep_out_path,
+            "digest": {"sha256": metadata_digest},
+            "annotations": {"version": "1.2.3"},
+        }
+    ]
+
+
+def test_get_dependencies_maps_env_only_output_paths_back_to_derivations(monkeypatch):
+    drv_path = "/nix/store/0aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-root.drv"
+    dep_out_basename = "2ccccccccccccccccccccccccccccccc-source"
+    dep_out_path = f"/nix/store/{dep_out_basename}"
+    digest = "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"
+
+    def fake_exec_cmd(cmd, **kwargs):
+        if cmd == [
+            "nix-store",
+            "--query",
+            "--references",
+            "--include-outputs",
+            drv_path,
+        ]:
+            return SimpleNamespace(stdout=f"{dep_out_path}\n")
+        if cmd[:4] == ["nix", "derivation", "show", "-r"]:
+            return SimpleNamespace(
+                stdout=json.dumps(
+                    {
+                        "derivations": {
+                            "1bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb-source.drv": {
+                                "name": "special-source",
+                                "outputs": {"out": {"method": "nar"}},
+                                "env": {
+                                    "out": dep_out_basename,
+                                    "version": "1.2.3",
+                                },
+                            }
+                        },
+                        "version": 4,
+                    }
+                )
+            )
+        raise AssertionError(f"unexpected command: {cmd} kwargs={kwargs}")
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fake_exec_cmd)
+    monkeypatch.setattr(
+        provenance_main,
+        "query_store_hashes",
+        lambda paths: ["sha256:1b8m03r63zqhnjf7l5wnldhh7c134ap5vpj0850ymkq1iyzicy5s"],
+    )
+
+    assert provenance_main.get_dependencies(drv_path) == [
+        {
+            "name": "special-source",
+            "uri": dep_out_path,
+            "digest": {"sha256": digest},
+            "annotations": {"version": "1.2.3"},
+        }
+    ]
+
+
+def test_get_subjects_falls_back_to_env_output_paths(monkeypatch):
+    output_path = "/custom/store/1bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb-nghttp2-1.68.1"
+    output_hash = "1b8m03r63zqhnjf7l5wnldhh7c134ap5vpj0850ymkq1iyzicy5s"
+    digest = "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"
+    real_exec_cmd = provenance_main.exec_cmd
+
+    def fake_exec_cmd(cmd, **kwargs):
+        if cmd == ["nix-store", "--query", "--hash", output_path]:
+            return SimpleNamespace(stdout=f"sha256:{output_hash}\n")
+        if cmd[:3] == ["nix", "hash", "convert"]:
+            return real_exec_cmd(cmd, **kwargs)
+        raise AssertionError(f"unexpected command: {cmd} kwargs={kwargs}")
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fake_exec_cmd)
+
+    assert provenance_main.get_subjects(
+        {"out": {"method": "nar"}},
+        env={"out": output_path},
+    ) == [
+        {
+            "name": "out",
+            "uri": output_path,
+            "digest": {"sha256": digest},
+        }
+    ]
+
+
+def test_get_subjects_prefers_derivation_hash_for_realized_flat_outputs(monkeypatch):
+    output_path = "/custom/store/1bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb-nghttp2-1.68.1"
+    output_hash = "sha256-ungWv48Bz+pBQUDeXa4iI7ADYaOWF3qctBD/YfIAFa0="
+    digest = "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"
+    real_exec_cmd = provenance_main.exec_cmd
+
+    def fake_exec_cmd(cmd, **kwargs):
+        if cmd[:3] == ["nix", "hash", "convert"]:
+            return real_exec_cmd(cmd, **kwargs)
+        raise AssertionError(f"unexpected command: {cmd} kwargs={kwargs}")
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fake_exec_cmd)
+
+    assert provenance_main.get_subjects(
+        {"out": {"method": "flat", "hash": output_hash}},
+        env={"out": output_path},
+    ) == [
+        {
+            "name": "out",
+            "uri": output_path,
+            "digest": {"sha256": digest},
+        }
+    ]
+
+
+def test_get_subjects_uses_derivation_hash_when_output_is_not_realized(monkeypatch):
+    output_path = "/custom/store/1bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb-nghttp2-1.68.1"
+    output_hash = "sha256-ungWv48Bz+pBQUDeXa4iI7ADYaOWF3qctBD/YfIAFa0="
+    digest = "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"
+    real_exec_cmd = provenance_main.exec_cmd
+
+    def fake_exec_cmd(cmd, **kwargs):
+        if cmd == ["nix-store", "--query", "--hash", output_path]:
+            return None
+        if cmd[:3] == ["nix", "hash", "convert"]:
+            return real_exec_cmd(cmd, **kwargs)
+        raise AssertionError(f"unexpected command: {cmd} kwargs={kwargs}")
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fake_exec_cmd)
+
+    assert provenance_main.get_subjects(
+        {"out": {"method": "nar", "hash": output_hash}},
+        env={"out": output_path},
+    ) == [
+        {
+            "name": "out",
+            "uri": output_path,
+            "digest": {"sha256": digest},
+        }
+    ]
+
+
+def test_get_subjects_supports_legacy_r_sha256_metadata(monkeypatch):
+    output_path = "/custom/store/1bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb-nghttp2-1.68.1"
+    digest = "77a94a83ccab42a68278ac5d3e340dcefecd736dd4feff1de71dec137b6b44ce"
+    real_exec_cmd = provenance_main.exec_cmd
+
+    def fake_exec_cmd(cmd, **kwargs):
+        if cmd[:3] == ["nix", "hash", "convert"]:
+            return real_exec_cmd(cmd, **kwargs)
+        raise AssertionError(f"unexpected command: {cmd} kwargs={kwargs}")
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fake_exec_cmd)
+
+    assert provenance_main.get_subjects(
+        {
+            "out": {
+                "hash": digest,
+                "hashAlgo": "r:sha256",
+            }
+        },
+        env={"out": output_path},
+    ) == [
+        {
+            "name": "out",
+            "uri": output_path,
+            "digest": {"sha256": digest},
+        }
+    ]
+
+
+def test_get_subjects_skips_unrealized_outputs_without_digest(monkeypatch):
+    output_path = "/custom/store/2ccccccccccccccccccccccccccccccc-nghttp2-doc"
+
+    def fake_exec_cmd(cmd, **_kwargs):
+        assert cmd == ["nix-store", "--query", "--hash", output_path]
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fake_exec_cmd)
+
+    assert not provenance_main.get_subjects(
+        {"out": {"method": "nar"}},
+        env={"out": output_path},
+    )
+
+
+def test_get_subjects_skip_only_missing_unrealized_outputs(monkeypatch):
+    output_path = "/custom/store/1bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb-nghttp2-1.68.1"
+    missing_path = "/custom/store/2ccccccccccccccccccccccccccccccc-nghttp2-doc"
+    output_hash = "1b8m03r63zqhnjf7l5wnldhh7c134ap5vpj0850ymkq1iyzicy5s"
+    digest = "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"
+    real_exec_cmd = provenance_main.exec_cmd
+
+    def fake_exec_cmd(cmd, **kwargs):
+        if cmd == ["nix-store", "--query", "--hash", output_path]:
+            return SimpleNamespace(stdout=f"sha256:{output_hash}\n")
+        if cmd == ["nix-store", "--query", "--hash", missing_path]:
+            return None
+        if cmd[:3] == ["nix", "hash", "convert"]:
+            return real_exec_cmd(cmd, **kwargs)
+        raise AssertionError(f"unexpected command: {cmd} kwargs={kwargs}")
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fake_exec_cmd)
+
+    assert provenance_main.get_subjects(
+        {"out": {"path": output_path}, "doc": {"path": missing_path}},
+    ) == [
+        {
+            "name": "out",
+            "uri": output_path,
+            "digest": {"sha256": digest},
+        }
+    ]
+
+
+def test_provenance_uses_store_path_hint_for_nix_2_33_outputs_without_path(monkeypatch):
+    target = "/custom/store/0aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-root.drv"
+    drv_basename = "0aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-root.drv"
+    out_basename = "1bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb-root"
+    output_hash = "sha256-ungWv48Bz+pBQUDeXa4iI7ADYaOWF3qctBD/YfIAFa0="
+    digest = "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"
+    real_exec_cmd = provenance_main.exec_cmd
+
+    def fake_exec_cmd(cmd, **kwargs):
+        if cmd[:3] == ["nix", "derivation", "show"]:
+            assert cmd[3] == target
+            return SimpleNamespace(
+                stdout=json.dumps(
+                    {
+                        "version": 4,
+                        "derivations": {
+                            drv_basename: {
+                                "name": "root",
+                                "outputs": {
+                                    "out": {
+                                        "method": "nar",
+                                        "hash": output_hash,
+                                    }
+                                },
+                                "env": {"out": out_basename},
+                            }
+                        },
+                    }
+                )
+            )
+        if cmd[:3] == ["nix", "hash", "convert"]:
+            return real_exec_cmd(cmd, **kwargs)
+        raise AssertionError(f"unexpected command: {cmd} kwargs={kwargs}")
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fake_exec_cmd)
+    monkeypatch.setattr(
+        provenance_main, "get_dependencies", lambda *_args, **_kwargs: []
+    )
+
+    metadata = provenance_main.BuildMeta("", "", "", "", "", "{}", "{}")
+    provenance = provenance_main.provenance(target, metadata)
+
+    assert provenance["subject"] == [
+        {
+            "name": "out",
+            "uri": f"/custom/store/{out_basename}",
+            "digest": {"sha256": digest},
+        }
+    ]
+
+
+def test_provenance_keeps_fixed_output_subjects_when_output_is_not_realized(
+    monkeypatch,
+):
+    target = "/custom/store/0aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-root.drv"
+    drv_basename = "0aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-root.drv"
+    out_basename = "1bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb-root"
+    output_hash = "sha256-ungWv48Bz+pBQUDeXa4iI7ADYaOWF3qctBD/YfIAFa0="
+    digest = "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"
+    real_exec_cmd = provenance_main.exec_cmd
+
+    def fake_exec_cmd(cmd, **kwargs):
+        if cmd[:3] == ["nix", "derivation", "show"]:
+            assert cmd[3] == target
+            return SimpleNamespace(
+                stdout=json.dumps(
+                    {
+                        "version": 4,
+                        "derivations": {
+                            drv_basename: {
+                                "name": "root",
+                                "outputs": {
+                                    "out": {"method": "nar", "hash": output_hash}
+                                },
+                                "env": {"out": out_basename},
+                            }
+                        },
+                    }
+                )
+            )
+        if cmd == ["nix-store", "--query", "--hash", f"/custom/store/{out_basename}"]:
+            return None
+        if cmd[:3] == ["nix", "hash", "convert"]:
+            return real_exec_cmd(cmd, **kwargs)
+        raise AssertionError(f"unexpected command: {cmd} kwargs={kwargs}")
+
+    monkeypatch.setattr(provenance_main, "exec_cmd", fake_exec_cmd)
+    monkeypatch.setattr(
+        provenance_main, "get_dependencies", lambda *_args, **_kwargs: []
+    )
+
+    metadata = provenance_main.BuildMeta("", "", "", "", "", "{}", "{}")
+    provenance = provenance_main.provenance(target, metadata)
+
+    assert provenance["subject"] == [
+        {
+            "name": "out",
+            "uri": f"/custom/store/{out_basename}",
+            "digest": {"sha256": digest},
+        }
+    ]
